% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.inapplicables.R
\name{apply.inapplicables}
\alias{apply.inapplicables}
\title{Apply inapplicable characters to a matrix.}
\usage{
apply.inapplicables(...)
}
\arguments{
\item{...}{Any additional arguments.}

\item{matrix}{A discrete morphological matrix.}

\item{inapplicables}{Optional, a vector of characters inapplicability source (either \code{"character"} or \code{"clade"}; see details). The length of this vector must be at maximum half the total number of characters.}

\item{invariant}{Whether to allow any invariant sites among the characters with inapplicable data.}
}
\description{
Apply inapplicable characters to discrete morphological matrix.
}
\details{
\itemize{
\item The \code{inapplicables} argument intakes a vector of character inapplicability source rendering a number of characters inapplicable using the following sources:
     \itemize{
         \item \code{"character"} draws inapplicable characters directly from the character matrix, ignoring the phylogeny (i.e. for a random character X, an other random character Y will have inappicable characters fro each character states 0 for character X).
         \item \code{"clade"} draws inapplicable characters from the phylogeny: it will randomly apply inapplicable characters states for some characters by randomly selecting clades from the provided tree.
     }
For example \code{inapplicables = c(rep("character", 2), rep("clade", 2))} will generate 4 characters with inapplicable data, two using previous characters and two other using random clades.

}
}
\examples{

}
\author{
Thomas Guillerme
}

