% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.tnt.R
\name{export.tnt}
\alias{export.tnt}
\title{Export a matrix to tnt.}
\usage{
export.tnt(matrix, chain.name, inapplicable = TRUE, run.n.read = FALSE)
}
\arguments{
\item{matrix}{A discrete morphological matrix.}

\item{chain.name}{A string of characters to be the tnt files chain name.}

\item{inapplicable}{\code{logical} whether to treat inapplicable as \code{-} (\code{TRUE}) or as \code{?} (\code{FALSE}).}

\item{run.n.read}{\code{logical} whether to run the matrix and read in the results (\code{default = FALSE}). For UNIX machines only: see details.}
}
\value{
Generates a tnt file out of the R environment. If \code{run.n.read = TRUE} also reads in the resulting trees and parsimony score.
}
\description{
Generates a tnt file that contains the matrix and the tnt instructions for running a tree.
}
\details{
The \code{run.n.read} arguments works by calling \code{system("tnt.command < chain.name.tnt")}. Make sure the \code{tnt.command} file is in the the \code{usr/bin/} or \code{usr/local/bin/} directory.
}
\examples{
## A random tree with 10 tips
tree <- rcoal(10)
## setting up the parameters
my_rates = c(rgamma, 1, 1) # A gamma rate distribution with of shape alpha = 0.5
my_substitutions = c(runif, 2, 2) # A fixed substitution rate of 2 (T/T ratio in HKY)

## Mk matrix (10*50)
matrixMk <- make.matrix(tree, characters = 50, model = "ER", rates = my_rates)

## Exporting to tnt
export.tnt(matrixMk, chain.name = "dummy")

## Tidying
remove.files("dummy.tnt")
}
\author{
Thomas Guillerme
}

