% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.contrast.matrix}
\alias{get.contrast.matrix}
\title{Generates a contrast matrix.}
\usage{
get.contrast.matrix(matrix)
}
\arguments{
\item{matrix}{a discrete character matrix.}
}
\description{
Creates a contrast matrix using the observed character states in an input matrix.
}
\author{
Thomas Guillerme
}

