% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeAlgo-package.R
\docType{package}
\name{IterativeAlgo-package}
\alias{IterativeAlgo}
\alias{IterativeAlgo-package}
\title{Inapplicable data stuff}
\description{
Utilities for inapplicable project
}
\author{
Thomas Guillerme <guillert@tcd.ie>
Maintainer: Thomas Guillerme
}

