% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrix.R
\name{make.matrix}
\alias{make.matrix}
\title{Generates a morphological matrix.}
\usage{
make.matrix(tree, characters, states = 1, model = "ER", rates, substitution,
  ..., invariant = FALSE, verbose = FALSE)
}
\arguments{
\item{tree}{A phylogenetic tree to use for generating the characters.}

\item{characters}{The number of morphological characters.}

\item{states}{A string of probabilities for the number of states for each characters (\code{default = 1}; i.e. 100\% binary state characters; see details).}

\item{model}{Either an implemented (\code{"ER"} or \code{"HKY"}; see details) or user defined model for generating one discrete morphological characters with at least the following arguments: \code{tree}, \code{states}, \code{rates}, \code{substitution}.}

\item{rates}{A function an it's parameters for the rates distribution (see details).}

\item{substitution}{A function an it's parameters for the substitutions distribution (see details).}

\item{...}{Any optional arguments to be passed to the model argument.}

\item{invariant}{Whether to allow any invariant sites.}

\item{verbose}{Whether to be verbose or not.}

\item{inapplicable}{Optional, a vector of characters inapplicability source (either \code{"character"} or \code{"clade"}; see details). The length of this vector must be at maximum half the total number of characters.}

\item{output}{Optional, an output file name for writing the matrix out of the \code{R} environement in \code{nexus} format.}
}
\description{
Generates a morphological matrix using \code{\link[ape]{rTraitDisc}} function with some optional inapplicable characters and allows to save it in various formats out of \code{R} environment.
}
\details{
\itemize{

\item The \code{model} arguments must be either a user's defined function for generating the discrete morphological characters (that intakes the states, rates and substitution arguments) or one of the two following:
     \itemize{
         \item \code{"ER"} uses the \code{ape::rTraitDisc} function with the \code{"ER"} model argument (= Mk model).
         \item \code{"HKY"} uses the \code{phyclust::gen.seq.HKY} function with \code{kappa} sampled from the \code{substitution} argument, \code{pi = runif(4)} (divided by \code{sum(runif(4))}), \code{rate.scale} sampled from the \code{rates} distribution and \code{L} being the number of \code{characters} and transforms the purines (A, G) into 0 and the pyrimidines (C, T) into 1.
     }

\item The \code{states} argument attributes a number of states to each character by using the given probability vector for each number of states starting from 2.
For example \code{states = c(0.7, 0.2, 0.1)} will generate 70% of characters with 2 states, 20% of characters with 3 states and 10% of characters with 4 states. 

\item The \code{rates} and \code{substitution} arguments attributes a distribution function and it's optional parameters to a model. For example \code{rates = c(runif, 1, 10)} attributes a uniform distribution between 1 and 10 for the rates distribution.


}
}
\examples{
## A random tree with 10 tips
tree <- rcoal(10)
## setting up the parameters
my_rates = c(rgamma, 1, 1) # A gamma rate distribution with of shape alpha = 0.5
my_substitutions = c(runif, 2, 2) # A fixed substitution rate of 2 (T/T ratio in HKY)

## A pure Mk matrix (10*50)
matrixMk <- make.matrix(tree, characters = 50, model = "ER", rates = my_rates)
## HKY binary (10*50)
matrixHKY <- make.matrix(tree, characters = 50, model="HKY", rates = my_rates, substitution = my_substitutiom)
}
\author{
Thomas Guillerme
}

