% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rtree.bd}
\alias{rtree.bd}
\title{Generates a birth death tree.}
\usage{
rtree.bd(n)
}
\arguments{
\item{n}{an integer giving the number of tips in the tree.}
}
\description{
Generates a birth death tree with a set number of taxa and a random birth and death rate.
}
\author{
Thomas Guillerme
}

