% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.matrix.R
\name{check.matrix}
\alias{check.matrix}
\title{Check morphological matrix consistency.}
\usage{
check.matrix(matrix, parismony = "fitch", first.tree = c(dist.hamming, NJ),
  orig.tree, distance = RF.dist, ...)
}
\arguments{
\item{matrix}{A discrete morphological matrix.}

\item{first.tree}{A list of functions to generate the first parsimonious tree (default = \code{c(\link[phangorn]{dist.hamming}, \link[phangorn]{NJ})}; see details).}

\item{orig.tree}{Optional, the input tree to measure the distance between the parsimony and the original tree.}

\item{distance}{Optional, if orig.tree is provide, the function to use for measuring distance between the trees (default = \code{link[phangorn]{RF.dist}}).}

\item{...}{Any additional arguments.}

\item{parsimony}{Either the parsimony algorithm to be passed to \code{\link[phangorn]{optim.parsimony}} or a parsimony function that can take a \code{\link[phangorn]{phyDat}} object as an input.}

\item{contrast.matrix}{An optional contrast matrix. By default, the function recognises any character states token as different apart from \code{?} that is treated as a joker character.}
}
\value{
Returns the parsimony score (using \code{\link[phangorn]{parsimony}}), the consistency and retention indices (using \code{\link[phangorn]{CI}} and \\code{\link[phangorn]{RI}}) from quick most parsimonious tree obtained from the matrix.
Can also return the topological distance from the original tree if provided.

#' @details
\itemize{
\item The \code{first.tree} argument must be a list of functions to be use in a cascade to transform the matrix (as a \code{\link[phangorn]{phyDat}} object) into a tree using the functions iteratively.
For example the default \code{c(\link[phangorn]{dist.hamming}, \link[phangorn]{NJ})} will apply the following to the matrix: \code{\link[phangorn]{NJ}(\link[phangorn]{dist.hamming}(matrix))}
}
}
\description{
Performs a quick and dirty checking of the phylogenetic signal in a morphological matrix using parismony.
}
\examples{
## Generating a random tree
random.tree <- rcoal(10)
## Generating a random matrix
random.matrix <- make.matrix(random.tree, characters = 50, model = "ER", rates = c(rgamma, 1, 1))

## Checking the matrix scores
check.matrix(random.matrix, orig.tree = random.tree)

}
\author{
Thomas Guillerme
}

